+++ WordPress/src/main/java/org/wordpress/android/ui/reader/views/compose/dropdown/JetpackDropdownMenuButton.kt
@@ -1,4 +1,4 @@
-package org.wordpress.android.ui.compose.components.menu.dropdown
+package org.wordpress.android.ui.reader.views.compose.dropdown
 
 import android.content.res.Configuration
 import androidx.compose.animation.animateContentSize
@@ -29,7 +29,7 @@ import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.IntSize
 import androidx.compose.ui.unit.dp
 import org.wordpress.android.R
-import org.wordpress.android.ui.compose.components.menu.dropdown.MenuElementData.Item
+import org.wordpress.android.ui.reader.views.compose.dropdown.MenuElementData.Item
 import org.wordpress.android.ui.compose.theme.AppThemeM3
 import org.wordpress.android.ui.compose.unit.Margin
 import org.wordpress.android.ui.compose.utils.uiStringText
