+++ WordPress/src/main/java/org/wordpress/android/ui/reader/views/compose/dropdown/JetpackDropdownMenuColors.kt
@@ -1,4 +1,4 @@
-package org.wordpress.android.ui.compose.components.menu.dropdown
+package org.wordpress.android.ui.reader.views.compose.dropdown
 
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material3.MaterialTheme
